+++ magma/compile.py
@@ -25,7 +25,7 @@ class CheckAnyMantleCircuits(DefinitionPass):
         self.has_mantle_circuit = False
 
     def __call__(self, definition):
-        if definition.debug_info.module is not None and \
+        if getattr(definition, "debug_info", False) and \
                 definition.debug_info.module.__name__.split(".")[0] == "mantle":
             self.has_mantle_circuit = True