+++ magma/syntax/sequential.py
@@ -7,6 +7,7 @@ from magma.debug import debug_info
 import functools
 import magma as m
 from magma.ssa import convert_tree_to_ssa
+from magma.config import get_debug_mode
 from collections import Counter
 
 
@@ -268,8 +269,8 @@ class SpecializeConstantInts(ast.NodeTransformer):
 
 
 def _sequential(defn_env: dict, cls):
-    if not inspect.isclass(cls):
-        raise ValueError("sequential decorator only works with classes")
+    # if not inspect.isclass(cls):
+    #     raise ValueError("sequential decorator only works with classes")
 
     initial_value_map = get_initial_value_map(cls.__init__, defn_env)
 
@@ -355,9 +356,10 @@ def _sequential(defn_env: dict, cls):
             ast.parse("from mantle import DefineRegister").body[0],
         ] + tree.body)
     circuit_def = ast_utils.compile_function_to_file(tree, cls.__name__, defn_env)
-    circuit_def.debug_info = debug_info(circuit_def.debug_info.filename,
-                                        circuit_def.debug_info.lineno,
-                                        inspect.getmodule(cls))
+    if get_debug_mode() and getattr(circuit_def, "debug_info", False):
+        circuit_def.debug_info = debug_info(circuit_def.debug_info.filename,
+                                            circuit_def.debug_info.lineno,
+                                            inspect.getmodule(cls))
 
     return circuit_def