+++ tests/test_circuit/test_inspect.py
@@ -1,5 +1,16 @@
 import magma as m
 
+def test_str_repr_core():
+    #Dont care what values they return as long as they retun non empty strings
+    assert str(m.Circuit)
+    assert repr(m.Circuit)
+    assert str(m.circuit.CircuitKind)
+    assert repr(m.circuit.CircuitKind)
+    assert str(m.CircuitType)
+    assert repr(m.CircuitType)
+    assert str(m.AnonymousCircuitType)
+    assert repr(m.AnonymousCircuitType)
+
 
 def test_str_repr():
     And2 = m.DeclareCircuit('And2', "I0", m.In(m.Bit), "I1", m.In(m.Bit),
@@ -14,12 +25,12 @@ def test_str_repr():
     print(repr(Logic2))
     assert repr(Logic2) == """\
 Logic2 = DefineCircuit("Logic2", "I0", In(Bit), "I1", In(Bit), "O", Out(Bit))
-XOr2_inst0 = XOr2()
 And2_inst0 = And2()
-wire(And2_inst0.O, XOr2_inst0.I0)
-wire(1, XOr2_inst0.I1)
+XOr2_inst0 = XOr2()
 wire(Logic2.I0, And2_inst0.I0)
 wire(Logic2.I1, And2_inst0.I1)
+wire(And2_inst0.O, XOr2_inst0.I0)
+wire(1, XOr2_inst0.I1)
 wire(XOr2_inst0.O, Logic2.O)
 EndCircuit()\
 """