+++ tests/test_verilog/test_from_file.py
@@ -1,6 +1,8 @@
 import magma as m
 import magma.testing
 import os
+import pytest
+
 
 def check_port(definition, port, type, direction):
     assert hasattr(definition, port)
@@ -76,7 +78,7 @@ def test_decl_list():
 
 def test_from_sv():
     file_path = os.path.dirname(__file__)
-    test_pe = m.DefineFromVerilogFile(os.path.join(file_path, "test_pe.sv"))[0]
+    test_pe = m.DefineFromVerilogFile(os.path.join(file_path, "test_pe.sv"), shallow=True)[0]
 
 
     if os.path.exists("build/test_pe.sv"):
@@ -92,3 +94,113 @@ def test_from_sv():
 
     assert m.testing.check_files_equal(__file__, "build/test_pe.sv",
                                        "test_pe.sv")
+
+
+def test_from_pad_inout():
+    file_path = os.path.dirname(__file__)
+    Pad = m.DeclareFromVerilogFile(os.path.join(file_path, "pad.v"))[0]
+
+    class Top(m.Circuit):
+        IO = ["pad", m.InOut(m.Bit)]
+        @classmethod
+        def definition(io):
+            pad = Pad()
+            m.wire(io.pad, pad.PAD)
+            for port in ["DS0", "DS1", "DS2", "I", "IE", "OEN", "PU", "PD", "ST", "SL", "RTE"]:
+                m.wire(0, getattr(pad, port))
+
+
+    m.compile("build/test_pad", Top, output="coreir-verilog")
+    assert m.testing.check_files_equal(__file__, "build/test_pad.v",
+                                       "gold/test_pad.v")
+
+
+def test_verilog_dependency():
+    [foo, bar] = m.DefineFromVerilog("""
+module foo(input I, output O);
+    assign O = I;
+endmodule
+
+module bar(input I, output O);
+    foo foo_inst(I, O);
+endmodule""")
+    top = m.DefineCircuit("top", "I", m.In(m.Bit), "O", m.Out(m.Bit))
+    bar_inst = bar()
+    m.wire(top.I, bar_inst.I)
+    m.wire(bar_inst.O, top.O)
+    m.EndDefine()
+    FILENAME = "test_verilog_dependency_top"
+    m.compile(f"build/{FILENAME}", top, output="coreir")
+    assert m.testing.check_files_equal(__file__, f"build/{FILENAME}.json",
+                                       f"gold/{FILENAME}.json")
+
+
+def test_verilog_dependency_out_of_order():
+    [foo, bar] = m.DefineFromVerilog("""
+module bar(input I, output O);
+    foo foo_inst(I, O);
+endmodule
+
+module foo(input I, output O);
+    assign O = I;
+endmodule""")
+    top = m.DefineCircuit("top", "I", m.In(m.Bit), "O", m.Out(m.Bit))
+    bar_inst = bar()
+    m.wire(top.I, bar_inst.I)
+    m.wire(bar_inst.O, top.O)
+    m.EndDefine()
+    FILENAME = "test_verilog_dependency_top"
+    m.compile(f"build/{FILENAME}", top, output="coreir")
+    assert m.testing.check_files_equal(__file__, f"build/{FILENAME}.json",
+                                       f"gold/{FILENAME}.json")
+
+
+def test_from_verilog_external_modules():
+    [foo] = m.DefineFromVerilog("""
+module foo(input I, output O);
+    assign O = I;
+endmodule""")
+    [bar] = m.DefineFromVerilog("""
+module bar(input I, output O);
+    foo foo_inst(I, O);
+endmodule""", external_modules={"foo": foo})
+    top = m.DefineCircuit("top", "I", m.In(m.Bit), "O", m.Out(m.Bit))
+    bar_inst = bar()
+    m.wire(top.I, bar_inst.I)
+    m.wire(bar_inst.O, top.O)
+    m.EndDefine()
+    FILENAME = "test_verilog_dependency_top"
+    m.compile(f"build/{FILENAME}", top, output="coreir")
+    assert m.testing.check_files_equal(__file__, f"build/{FILENAME}.json",
+                                       f"gold/{FILENAME}.json")
+
+
+def test_from_verilog_external_modules_missing():
+    with pytest.raises(Exception) as pytest_e:
+        m.DefineFromVerilog("""
+module bar(input I, output O);
+    foo foo_inst(I, O);
+endmodule""")
+        assert False
+    assert pytest_e.type is KeyError
+    assert pytest_e.value.args == ("foo",)
+
+
+def test_from_verilog_external_modules_duplicate():
+    with pytest.raises(Exception) as pytest_e:
+        [foo] = m.DefineFromVerilog("""
+module foo(input I, output O);
+    assign O = I;
+endmodule""")
+        [bar] = m.DefineFromVerilog("""
+module foo(input I, output O);
+    assign O = I;
+endmodule
+
+module bar(input I, output O);
+    foo foo_inst(I, O);
+endmodule""", external_modules={"foo": foo})
+        assert False
+    assert pytest_e.type is Exception
+    assert pytest_e.value.args == \
+        ("Modules defined in both external_modules and in parsed verilog: {'foo'}",)  # nopep8