+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.4] - 2019-04-25
+### Fixed
+- Fixed bug in coreir backend's handling of inout ports
+
+## [1.0.2] - 2019-04-24
+### Fixed
+- Fixed bug in `BFloatKind` to remove second `__getitem__` definition
+
 ## [1.0.0] - 2019-03-20
 ### Changed
 https://github.com/phanrahan/magma/pull/369